Fix bytesToStr on CalcBytesType
Byte array maps to a normal array in JavaScript. However, method
subarray() is available only on TypedArrays, not on regular Array.

That means that createStringFromArray doesn't work on CalcBytesType.
Method slice() should work on both Arrays and TypedArrays, but it's
safer to use subarray() on TypedArrays, because it doesn't make a
memory copy (it uses to the same ArrayBuffer according to the docs)
and is more compatible across the browsers.